rename: CampDetails.tsx -> index.tsx
