+++ src/pages/CampDetail/index.tsx
